feat: store 모듈 분리
